+++ x/evm/handler.go
@@ -74,10 +74,7 @@ func handleMsgEthereumTx(ctx sdk.Context, k Keeper, msg types.MsgEthereumTx) (*s
 	k.Bloom.Or(k.Bloom, executionResult.Bloom)
 
 	// update transaction logs in KVStore
-	err = k.SetTransactionLogs(ctx, executionResult.Logs, txHash)
-	if err != nil {
-		return nil, err
-	}
+	k.SetTransactionLogs(ctx, txHash, executionResult.Logs)
 
 	// log successful execution
 	k.Logger(ctx).Info(executionResult.Result.Log)
@@ -150,10 +147,7 @@ func handleMsgEthermint(ctx sdk.Context, k Keeper, msg types.MsgEthermint) (*sdk
 	k.Bloom.Or(k.Bloom, executionResult.Bloom)
 
 	// update transaction logs in KVStore
-	err = k.SetTransactionLogs(ctx, executionResult.Logs, txHash)
-	if err != nil {
-		return nil, err
-	}
+	k.SetTransactionLogs(ctx, txHash, executionResult.Logs)
 
 	// log successful execution
 	k.Logger(ctx).Info(executionResult.Result.Log)
