+++ x/evm/types/genesis.go
@@ -1,12 +1,14 @@
 package types
 
 import (
+	"bytes"
 	"errors"
 	"math/big"
 
-	"github.com/cosmos/ethermint/types"
 	ethcmn "github.com/ethereum/go-ethereum/common"
 )
+
+var zeroAddrBytes = []byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0}
 
 type (
 	// GenesisState defines the application's genesis state. It contains all the
@@ -15,31 +17,49 @@ type (
 		Accounts []GenesisAccount `json:"accounts"`
 	}
 
+	// GenesisStorage represents the GenesisAccount Storage map as single key value
+	// pairs. This is to prevent non determinism at genesis initialization or export.
+	GenesisStorage struct {
+		Key   ethcmn.Hash `json:"key"`
+		Value ethcmn.Hash `json:"value"`
+	}
+
 	// GenesisAccount defines an account to be initialized in the genesis state.
+	// Its main difference between with Geth's GenesisAccount is that it uses a custom
+	// storage type and that it doesn't contain the private key field.
 	GenesisAccount struct {
-		Address ethcmn.Address `json:"address"`
-		Balance *big.Int       `json:"balance"`
-		Code    []byte         `json:"code,omitempty"`
-		Storage types.Storage  `json:"storage,omitempty"`
+		Address ethcmn.Address   `json:"address"`
+		Balance *big.Int         `json:"balance"`
+		Code    []byte           `json:"code,omitempty"`
+		Storage []GenesisStorage `json:"storage,omitempty"`
 	}
 )
 
-// ValidateGenesis validates evm genesis config
-func ValidateGenesis(data GenesisState) error {
-	for _, acct := range data.Accounts {
-		if len(acct.Address.Bytes()) == 0 {
-			return errors.New("invalid GenesisAccount: address cannot be empty")
-		}
-		if acct.Balance == nil {
-			return errors.New("invalid GenesisAccount: balance cannot be empty")
-		}
+// NewGenesisStorage creates a new GenesisStorage instance
+func NewGenesisStorage(key, value ethcmn.Hash) GenesisStorage {
+	return GenesisStorage{
+		Key:   key,
+		Value: value,
 	}
-	return nil
 }
 
-// DefaultGenesisState sets default evm genesis config
+// DefaultGenesisState sets default evm genesis state with empty accounts.
 func DefaultGenesisState() GenesisState {
 	return GenesisState{
 		Accounts: []GenesisAccount{},
 	}
+}
+
+// Validate performs basic genesis state validation returning an error upon any
+// failure.
+func (gs GenesisState) Validate() error {
+	for _, acc := range gs.Accounts {
+		if bytes.Equal(acc.Address.Bytes(), zeroAddrBytes) {
+			return errors.New("invalid GenesisAccount: address cannot be empty")
+		}
+		if acc.Balance == nil {
+			return errors.New("invalid GenesisAccount: balance cannot be empty")
+		}
+	}
+	return nil
 }
